fix fe ut
